New models + Actor WIP

+++ silo-core/src/models/subject.rs
@@ -1,10 +1,10 @@
 /// Contains information about a single subject for analysis.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Subject {
     /// The subject's unique ID.
-    pub id: i32,
+    pub id: i16,
     /// The ID of the group the subject belongs to.
-    pub group_id: i32,
+    pub group_id: i16,
     /// The subject's age.
     pub age: i8,
     /// The subject's length of stay.
